extract_var

Extract variables from pulled variables

.extract_var(
  unique_positions,
  env_stack,
  env_var,
  .fill_gaps,
  .buffer = NULL,
  verbose = TRUE
)

Arguments

unique_positions
dataframe with unique coordinates and dates from where to extract daily environmental data

env_stack
a raster stack object with daily spatial environmental layers associated with detection dates (output from .pull_var function)

env_var
name of extracted variable (brought in from extractEnv function)

.fill_gaps
Fill spatial gaps within each layer using a distance buffer

.buffer
radius of buffer (in m) around each detection from which environmental variables should be extracted from. A median value of pixels that fall within the buffer will be used if .fill_gaps = TRUE

verbose
should function provide details of what operation is being conducted. Uses parameter from parent function extractEnv

Value

a dataframe with unique position, unique dates and extracted environmental variables

Details

Internal function to conduct raster extraction of environmental variables from downloaded raster stack (env_stack) and detection positions (unique_positions)