Find closest mooring

Links the QC Detection Dataset to the Nearest IMOS Mooring ID

getDistance(
  trackingData,
  moorLocations,
  X = "receiver_deployment_longitude",
  Y = "receiver_deployment_latitude",
  datetime = "detection_datetime"
)

Arguments

trackingData

dataframe containing acoustic detection data in IMOS QC format

moorLocations

dataframe containing the locations of IMOS moorings

X

name of column with X coordinate or longitude (EPSG 4326)

Y

name of column with Y coordinate or latitude (EPSG 4326)

datetime

name of column with date time stamp (Coordinated Universal Time; UTC)

Value

The trackingData dataframe with the nearest moor_site_code to each acoustic detection