Table of all available IMOS variables accessible from remora package

Summary of available IMOS environmental variable to append to detection data.

imos_variables(variable = NULL)

Arguments

variable
(optional) name of a specific variable you are interested in. Default of NULL provides details for all available datasets.

Value

a formatted table with details of available variables accessible through the remora package

Details

This function helps users identify what quality controlled environmental variables can be accessed and appended to their detection data using the extractEnv of mooringTable functions.

Examples

## Identify all available variables
imos_variables()
#> <table class=" lightable-paper" style='font-family: "Arial Narrow", arial, helvetica, sans-serif; width: auto !important; margin-left: auto; margin-right: auto;'>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Variable </th>
#>    <th style="text-align:left;"> Platform </th>
#>    <th style="text-align:left;"> Temporal resolution </th>
#>    <th style="text-align:left;"> Units </th>
#>    <th style="text-align:left;"> Function to use </th>
#>    <th style="text-align:left;"> Description </th>
#>    <th style="text-align:left;"> Source </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> bathy </td>
#>    <td style="text-align:left;"> Composite raster product </td>
#>    <td style="text-align:left;"> - </td>
#>    <td style="text-align:left;"> meters </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> Australian Bathymetry and Topography Grid. 250 m resolution. </td>
#>    <td style="text-align:left;"> Geosciences Australia </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> dist_to_land </td>
#>    <td style="text-align:left;"> Raster product </td>
#>    <td style="text-align:left;"> - </td>
#>    <td style="text-align:left;"> kilometers </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> Distance from nearest shoreline (in km). Derived from the high-resolution Open Street Map shoreline product. </td>
#>    <td style="text-align:left;"> This package </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> rs_sst </td>
#>    <td style="text-align:left;"> Satellite-derived raster product </td>
#>    <td style="text-align:left;"> daily (2002-07-04 - present) </td>
#>    <td style="text-align:left;"> degrees Celcius </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> 1-day multi-swath multi-sensor (L3S) remotely sensed sea surface temperature (degrees Celcius) at 2 km resolution. Derived from the Group for High Resolution Sea Surface Temperature (GHRSST) </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> rs_sst_interpolated </td>
#>    <td style="text-align:left;"> Raster product </td>
#>    <td style="text-align:left;"> daily (2006-06-12 - present) </td>
#>    <td style="text-align:left;"> degrees Celcius </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> 1-day interpolated remotely sensed sea surface temperature (degrees Celcius) at 9 km resolution. Derived from the Regional Australian Multi-Sensor Sea surface temperature Analysis (RAMSSA, Beggs et al. 2010) system as part of the BLUElink Ocean Forecasting Australia project </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> rs_chl </td>
#>    <td style="text-align:left;"> Satellite-derived raster product </td>
#>    <td style="text-align:left;"> daily (2002-07-04 - present) </td>
#>    <td style="text-align:left;"> mg.m-3 </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> Remotely sensed chlorophyll-a concentration (OC3 model). Derived from the MODIS Aqua satellite mission. Multi-spectral measurements are used to infer the concentration of chlorophyll-a, most typically due to phytoplankton, present in the water (mg.m-3). </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> rs_current </td>
#>    <td style="text-align:left;"> Composite raster product </td>
#>    <td style="text-align:left;"> daily (1993-01-01 - present) </td>
#>    <td style="text-align:left;"> ms-1; degrees </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> Gridded (adjusted) sea level anomaly (GSLA), surface geostrophic velocity in the east-west (UCUR) and north-south (VCUR) directions for the Australasian region derived from the IMOS Ocean Current project. Two additional variables are calculated: surface current velocity (ms-1) and bearing (degrees). </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> rs_salinity </td>
#>    <td style="text-align:left;"> Satellite-derived raster product </td>
#>    <td style="text-align:left;"> weekly (2011-08-25 - 2015-06-07) </td>
#>    <td style="text-align:left;"> psu </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> 7-day composite remotely sensed salinity. Derived from the NASA Aquarius satellite mission (psu). </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> rs_turbidity </td>
#>    <td style="text-align:left;"> Satellite-derived raster product </td>
#>    <td style="text-align:left;"> daily (2002-07-04 - present) </td>
#>    <td style="text-align:left;"> m-1 </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> Diffuse attenuation coefficient at 490 nm (K490) indicates the turbidity of the water column (m-1). The value of K490 represents the rate which light at 490 nm is attenuated with depth. For example a K490 of 0.1/meter means that light intensity will be reduced one natural log within 10 meters of water. Thus, for a K490 of 0.1, one attenuation length is 10 meters. Higher K490 value means smaller attenuation depth, and lower clarity of ocean water. </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> rs_npp </td>
#>    <td style="text-align:left;"> Satellite-derived raster product </td>
#>    <td style="text-align:left;"> daily (2002-07-04 - present) </td>
#>    <td style="text-align:left;"> mgC.m_2.day-1 </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> Net primary productivity (OC3 model and Eppley-VGPM algorithm). Modelled product used to compute an estimate of the Net Primary Productivity (NPP). The model used is based on the standard vertically generalised production model (VGPM). The VGPM is a "chlorophyll-based" model that estimates net primary production from chlorophyll using a temperature-dependent description of chlorophyll-specific photosynthetic efficiency. For the VGPM, net primary production is a function of chlorophyll, available light, and the photosynthetic efficiency. The only difference between the Standard VGPM and the Eppley-VGPM is the temperature-dependent description of photosynthetic efficiencies, with the Eppley approach using an exponential function to account for variation in photosynthetic efficiencies due to photoacclimation. </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> moor_sea_temp </td>
#>    <td style="text-align:left;"> Fixed sub-surface moorings </td>
#>    <td style="text-align:left;"> hourly </td>
#>    <td style="text-align:left;"> degrees Celcius </td>
#>    <td style="text-align:left;"> extractMoor() </td>
#>    <td style="text-align:left;width: 30em; "> Depth-integrated in-situ, hourly time-series measurements of sea temperature (degrees Celcius) at fixed mooring locations </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> moor_psal </td>
#>    <td style="text-align:left;"> Fixed sub-surface moorings </td>
#>    <td style="text-align:left;"> hourly </td>
#>    <td style="text-align:left;"> psu </td>
#>    <td style="text-align:left;"> extractMoor() </td>
#>    <td style="text-align:left;width: 30em; "> Depth-integrated in-situ, hourly time-series measurements of salinity (psu) at fixed mooring locations </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> moor_ucur </td>
#>    <td style="text-align:left;"> Fixed sub-surface moorings </td>
#>    <td style="text-align:left;"> hourly </td>
#>    <td style="text-align:left;"> ms-1 </td>
#>    <td style="text-align:left;"> extractMoor() </td>
#>    <td style="text-align:left;width: 30em; "> Depth-integrated in-situ, hourly time-series measurements of subsurface geostrophic current velocity in the east-west direction (ms-1) at fixed mooring locations </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> moor_vcur </td>
#>    <td style="text-align:left;"> Fixed sub-surface moorings </td>
#>    <td style="text-align:left;"> hourly </td>
#>    <td style="text-align:left;"> ms-1 </td>
#>    <td style="text-align:left;"> extractMoor() </td>
#>    <td style="text-align:left;width: 30em; "> Depth-integrated in-situ, hourly time-series measurements of subsurface geostrophic current velocity in the north-south direction (ms-1) at fixed mooring locations </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#> </tbody>
#> </table>

## If there is a specific variable you are interested in
imos_variables(variable = "rs_sst_interpolated")
#> <table class=" lightable-paper" style='font-family: "Arial Narrow", arial, helvetica, sans-serif; width: auto !important; margin-left: auto; margin-right: auto;'>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Variable </th>
#>    <th style="text-align:left;"> Platform </th>
#>    <th style="text-align:left;"> Temporal resolution </th>
#>    <th style="text-align:left;"> Units </th>
#>    <th style="text-align:left;"> Function to use </th>
#>    <th style="text-align:left;"> Description </th>
#>    <th style="text-align:left;"> Source </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;font-weight: bold;"> rs_sst_interpolated </td>
#>    <td style="text-align:left;"> Raster product </td>
#>    <td style="text-align:left;"> daily (2006-06-12 - present) </td>
#>    <td style="text-align:left;"> degrees Celcius </td>
#>    <td style="text-align:left;"> extractEnv() </td>
#>    <td style="text-align:left;width: 30em; "> 1-day interpolated remotely sensed sea surface temperature (degrees Celcius) at 9 km resolution. Derived from the Regional Australian Multi-Sensor Sea surface temperature Analysis (RAMSSA, Beggs et al. 2010) system as part of the BLUElink Ocean Forecasting Australia project </td>
#>    <td style="text-align:left;"> IMOS </td>
#>   </tr>
#> </tbody>
#> </table>