find least-cost distance between two locations

calculates least-cost distance between two receiver locations if a land mass lies between them. Only applied on tags detected more than once

shortest_dist2(position, inst, raster = NULL, tr)

Arguments

position

locations of deployment and subsequent detections as a data.frame of longitude, latitude

inst

installation names of detection (receiver) locations

raster

raster of telemetry region coastline, if NULL then the default Aust_sr SpatRaster is used

tr

transition matrix for calculating least-cost distances (ie. produced via transition, with geoCorrection applied)

Value

a 1-column matrix of shortest distances

Details

...