write a QCsummary.csv file and/or QC processed .csv files for each tag deployment

writes a single QCsummary.csv file and/or tag-deployment-specific temporal_outcome's of quality-controlled detections to .csv files

writeQC(x, path = NULL, summary = TRUE, csv = TRUE)

Arguments

x

a nested tibble with class remora_QC generated by runQC

path

path to write QC'd files to. If NULL the files are written to the working directory

summary

whether to generate a summary .csv file. If TRUE then QCsummary.csv is written to the specified path

csv

whether to generate the individual .csv files. If TRUE (default) then each file is written to the specified path

Value

.csv files and/or a QCsummary.csv file are written to the specified path

Details

takes a remora_QC nested tibble and writes .csv files corresponding to each row of the nested tibble. A summary of the QC process is written to QCsummary.csv. The summary contains the following information for each QC'd tag (see references for more details):

total_detectionsthe number of raw detections

detections_before_deploymentthe number of detections recorded prior to deployment date in metadata

invalid_deployment_locationlogical indicating whether deployment longitude,latitude is valid

detections_outside_species_rangethe number of detections that occurred outside the species expert distribution

valid_detectionsthe number of detections that passed the QC process (QC flags 1 and 2)

tracking_duration_daysthe number of days between deployment and the last valid detection

invalid_velocitythe number of detections associated with implausible travel speeds

References

Hoenner X et al. (2018) Australia’s continental-scale acoustic tracking database and its automated quality control process. Sci Data 5, 170206 https://doi.org/10.1038/sdata.2017.206

Examples

if (FALSE) {
## example QC'd data
data(TownsvilleReefQC)
## write QC output
writeQC(TownsvilleReefQC, summary = TRUE)
}